refactor(injector): add exec() method
- Remove injectorFilter from plugin dir
- Move cache into Injector Class
- Remove decache
