+++ lib/hexo/index.js
@@ -25,7 +25,6 @@ const defaultConfig = require('./default_config');
 const loadDatabase = require('./load_database');
 const multiConfigPath = require('./multi_config_path');
 const { sync } = require('resolve');
-const injectorFilter = require('../plugins/injector');
 const { inherits } = require('util');
 const { deepMerge, full_url_for } = require('hexo-util');
 
@@ -57,7 +56,7 @@ const createLoadThemeRoute = function(generatorResult, locals, ctx) {
       if (view) {
         log.debug(`Rendering HTML ${name}: ${magenta(path)}`);
         return view.render(locals)
-          .then(result => injectorFilter.call(ctx, result, locals))
+          .then(result => ctx.extend.injector.exec(result, locals))
           .then(result => ctx.execFilter('after_route_render', result, {
             context: ctx,
             args: [locals]
@@ -225,6 +224,7 @@ class Hexo {
     require('../plugins/filter')(this);
     require('../plugins/generator')(this);
     require('../plugins/helper')(this);
+    require('../plugins/injector')(this);
     require('../plugins/processor')(this);
     require('../plugins/renderer')(this);
     require('../plugins/tag')(this);
