+++ package.json
@@ -71,7 +71,6 @@
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
     "cheerio": "0.22.0",
-    "decache": "^4.5.1",
     "eslint": "^6.0.1",
     "eslint-config-hexo": "^4.1.0",
     "hexo-renderer-marked": "^2.0.0",
