+++ test/scripts/extend/injector.js
@@ -1,6 +1,18 @@
 'use strict';
 
 describe('Injector', () => {
+  const content = [
+    '<!DOCTYPE html>',
+    '<html lang="en">',
+    '<head id="head"><title>Test</title>',
+    '</head>',
+    '<body id="body">',
+    '<div></div>',
+    '<p></p>',
+    '</body>',
+    '</html>'
+  ].join('');
+
   const Injector = require('../../../lib/extend/injector');
 
   it('register() - entry is required', () => {
@@ -80,5 +92,146 @@ describe('Injector', () => {
 
     i.getText('body_end', 'home').should.eql(str);
     i.getText('body_end').should.eql('');
+  });
+
+  it('exec() - default', () => {
+    const i = new Injector();
+    const result = i.exec(content);
+    result.should.contain('<head id="head"><title>Test</title></head>');
+    result.should.contain('<body id="body"><div></div><p></p></body>');
+  });
+
+  it('exec() - default', () => {
+    const i = new Injector();
+    const result = i.exec(content);
+    result.should.contain('<head id="head"><title>Test</title></head>');
+    result.should.contain('<body id="body"><div></div><p></p></body>');
+  });
+
+  it('exec() - insert code', () => {
+    const i = new Injector();
+
+    i.register('head_begin', '<!-- Powered by Hexo -->');
+    i.register('head_end', '<link href="prism.css" rel="stylesheet">');
+    i.register('head_end', '<link href="prism-linenumber.css" rel="stylesheet">');
+    i.register('body_begin', '<script>window.Prism = window.Prism || {}; window.Prism.manual = true;</script>');
+    i.register('body_end', '<script src="prism.js"></script>');
+
+    const result = i.exec(content);
+
+    result.should.contain('<head id="head"><!-- hexo injector head_begin start --><!-- Powered by Hexo --><!-- hexo injector head_begin end -->');
+    result.should.contain('<!-- hexo injector head_end start --><link href="prism.css" rel="stylesheet"><link href="prism-linenumber.css" rel="stylesheet"><!-- hexo injector head_end end --></head>');
+    result.should.contain('<body id="body"><!-- hexo injector body_begin start --><script>window.Prism = window.Prism || {}; window.Prism.manual = true;</script><!-- hexo injector body_begin end -->');
+    result.should.contain('<!-- hexo injector body_end start --><script src="prism.js"></script><!-- hexo injector body_end end --></body>');
+  });
+
+  it('exec() - no duplicate insert', () => {
+    const content = [
+      '<!DOCTYPE html>',
+      '<html lang="en">',
+      '<head id="head"><!-- hexo injector head_begin start --><!-- hexo injector head_begin end -->',
+      '<title>Test</title>',
+      '<!-- hexo injector head_end start --><link href="prism.css" rel="stylesheet"></head>',
+      '<body id="body"><!-- hexo injector body_begin start --><!-- hexo injector body_begin end -->',
+      '<div></div>',
+      '<p></p>',
+      '<!-- hexo injector body_end start --><script src="prism.js"></script><!-- hexo injector body_end end --></body>',
+      '</html>'
+    ].join('');
+
+    const i = new Injector();
+
+    i.register('head_begin', '<!-- Powered by Hexo -->');
+    i.register('head_end', '<link href="prism.css" rel="stylesheet">');
+    i.register('head_end', '<link href="prism-linenumber.css" rel="stylesheet">');
+    i.register('body_begin', '<script>window.Prism = window.Prism || {}; window.Prism.manual = true;</script>');
+    i.register('body_end', '<script src="prism.js"></script>');
+
+    const result = i.exec(content);
+
+    result.should.contain('<head id="head"><!-- hexo injector head_begin start --><!-- hexo injector head_begin end -->');
+    result.should.contain('<!-- hexo injector head_end start --><link href="prism.css" rel="stylesheet"></head>');
+    result.should.contain('<body id="body"><!-- hexo injector body_begin start --><!-- hexo injector body_begin end -->');
+    result.should.contain('<!-- hexo injector body_end start --><script src="prism.js"></script><!-- hexo injector body_end end --></body>');
+  });
+
+  it('exec() - multi-line head & body', () => {
+    const content = [
+      '<!DOCTYPE html>',
+      '<html lang="en">',
+      '<head id="head"><title>Test</title>',
+      '</head>',
+      '<body id="body">',
+      '<div></div>',
+      '<p></p>',
+      '</body>',
+      '</html>'
+    ].join('\n');
+
+    const i = new Injector();
+
+    i.register('head_begin', '<!-- Powered by Hexo -->');
+    i.register('head_end', '<link href="prism.css" rel="stylesheet">');
+    i.register('head_end', '<link href="prism-linenumber.css" rel="stylesheet">');
+    i.register('body_begin', '<script>window.Prism = window.Prism || {}; window.Prism.manual = true;</script>');
+    i.register('body_end', '<script src="prism.js"></script>');
+
+    const result = i.exec(content);
+
+    result.should.contain('<head id="head"><!-- hexo injector head_begin start --><!-- Powered by Hexo --><!-- hexo injector head_begin end -->');
+    result.should.contain('<!-- hexo injector head_end start --><link href="prism.css" rel="stylesheet"><link href="prism-linenumber.css" rel="stylesheet"><!-- hexo injector head_end end --></head>');
+    result.should.contain('<body id="body"><!-- hexo injector body_begin start --><script>window.Prism = window.Prism || {}; window.Prism.manual = true;</script><!-- hexo injector body_begin end -->');
+    result.should.contain('<!-- hexo injector body_end start --><script src="prism.js"></script><!-- hexo injector body_end end --></body>');
+  });
+
+  it('exec() - inject on specific page', () => {
+    const content = [
+      '<!DOCTYPE html>',
+      '<html lang="en">',
+      '<head id="head"><title>Test</title>',
+      '</head>',
+      '<body id="body">',
+      '<div></div>',
+      '<p></p>',
+      '</body>',
+      '</html>'
+    ].join('\n');
+
+    const i = new Injector();
+
+    i.register('head_begin', '<!-- head_begin_default -->');
+    i.register('head_begin', '<!-- head_begin_home -->', 'home');
+    i.register('head_begin', '<!-- head_begin_post -->', 'post');
+    i.register('head_begin', '<!-- head_begin_page -->', 'page');
+    i.register('head_begin', '<!-- head_begin_archive -->', 'archive');
+    i.register('head_begin', '<!-- head_begin_category -->', 'category');
+    i.register('head_begin', '<!-- head_begin_tag -->', 'tag');
+
+    const result1 = i.exec(content, { page: {} });
+    const result2 = i.exec(content, { page: { __index: true } });
+    const result3 = i.exec(content, { page: { __post: true } });
+    const result4 = i.exec(content, { page: { __page: true } });
+    const result5 = i.exec(content, { page: { archive: true } });
+    const result6 = i.exec(content, { page: { category: true } });
+    const result7 = i.exec(content, { page: { tag: true } });
+
+    // home
+    result1.should.not.contain('<!-- head_begin_home -->');
+    result2.should.contain('<!-- head_begin_home --><!-- head_begin_default -->');
+    // post
+    result1.should.not.contain('<!-- head_begin_post -->');
+    result3.should.contain('<!-- head_begin_post --><!-- head_begin_default -->');
+    // page
+    result1.should.not.contain('<!-- head_begin_page -->');
+    result4.should.contain('<!-- head_begin_page --><!-- head_begin_default -->');
+    // archive
+    result1.should.not.contain('<!-- head_begin_archive -->');
+    result5.should.contain('<!-- head_begin_archive --><!-- head_begin_default -->');
+    // category
+    result1.should.not.contain('<!-- head_begin_category -->');
+    result6.should.contain('<!-- head_begin_category --><!-- head_begin_default -->');
+    // tag
+    result1.should.not.contain('<!-- head_begin_tag -->');
+    result7.should.contain('<!-- head_begin_tag --><!-- head_begin_default -->');
   });
 });
